Sharpen the pickle tests and file the entry with the type
