+++ CHANGELOG.md
@@ -25,7 +25,7 @@ All notable changes to pylj are recorded here. The format follows
 - `MDSimulation` and `MCSimulation` refuse an initial configuration whose pair energy is not finite or stores more than `simulation.INITIAL_ENERGY_LIMIT` (ten) k_B T per atom, as an overlapping lattice does; the message gives the stored energy per atom in k_B T.
 - `MDSimulation` and `MCSimulation` refuse a pair potential whose energy at the cut-off is not finite or is larger in magnitude than k_B T, since the cut-off assumes the interaction has died away there; parameters in the wrong units are one way to trip it.
 - `placement_temperature` on `MDSimulation.initialise` and `MCSimulation.initialise`: the temperature of the Metropolis acceptance used to place an initial configuration, by default the run temperature.
-- `pylj.model.Model`, the species and the potential between each pair of them, validated when it is built and read-only after; `Model.single(species, potential)` builds the one-species case, and `Model.potential(one, other)` looks a pair up in either order.
+- `pylj.model.Model`, the species and the potential between each pair of them, validated when it is built and read-only after; `Model.single(species, potential)` builds the one-species case, and `Model.potential(one, other)` looks a pair up in either order. A model, and any simulation holding one, can be pickled and deep-copied, so a temperature scan can be handed to `multiprocessing` and a finished run saved to disk.
 - `LennardJones`, `Buckingham` and `SquareWell` print as the constructor call that built them, so a model shows its parameters in a notebook.
 - `placement.place_triangular` and `init_conf='triangular'`, which start a simulation from the triangular lattice a two-dimensional solid settles into, with `max_strain` on `place` and on both `initialise` methods setting how far the fitted lattice may sit from `placement.TRIANGULAR_RATIO`, as a fraction of that ratio, up to `placement.MOST_STRAIN`.
 - `pylj.scattering`, the wavevectors commensurate with a box and the structure factor of a configuration at them.
@@ -67,7 +67,6 @@ All notable changes to pylj are recorded here. The format follows
 
 ### Fixed
 
-- `Model`, and the simulations holding one, can be pickled and deep-copied, so a temperature scan can be handed to `multiprocessing` and a finished run saved to disk.
 - `Energy` and `Phase` viewers crashed under NumPy 2.
 - `Interactions`, `Phase` and `Scattering` crashed if built before the first sample.
 - Two-type systems drew the other type's atoms at the origin.
