+++ docs/source/simulations.md
@@ -18,7 +18,7 @@ simulation = MDSimulation.initialise(
 )
 ```
 
-`model`, a `Model` of the species and the potential between each pair of them, is the one positional argument; the rest are given by keyword. Three are required: `number_of_atoms`; `temperature`, in kelvin; and `box`, the side of the square periodic box in Angstrom. `init_conf` selects the starting positions: `"square"` places the atoms on a square lattice, `"triangular"` on a triangular one, and `"metropolis"` inserts them one at a time by Metropolis acceptance at `placement_temperature`, which defaults to `temperature`. A triangular lattice fills the box, so the number of atoms has to be a number of columns times an even number of rows. Each row is offset along x by half a column spacing from the row below, and with an odd number of rows the last row and the first meet unoffset across the periodic boundary, which breaks the lattice; 195 atoms, 13 columns by 15 rows, is refused for that reason rather than for its strain. Dividing the columns by the rows has to give something near `sqrt(3) / 2`, and `max_strain` is how far away it may sit, as a fraction of that ratio. At the default of 0.05 the counts up to 300 are 30, 56, 90, 120, 168, 224, 270 and 288, and any other count raises `ValueError` naming ones that fit. `timestep` is in seconds and applies to molecular dynamics only. `cut_off` is in Angstrom and defaults to 15 or half the box, whichever is smaller; it may not exceed half the box. `seed` seeds `simulation.rng`, which draws the initial velocities and the Monte Carlo moves.
+`model`, a `Model` of the species and the potential between each pair of them, is the one positional argument; the rest are given by keyword. Three are required: `number_of_atoms`; `temperature`, in kelvin; and `box`, the side of the square periodic box in Angstrom. `init_conf` selects the starting positions: `"square"` places the atoms on a square lattice, `"triangular"` on a triangular one, and `"metropolis"` inserts them one at a time by Metropolis acceptance at `placement_temperature`, which defaults to `temperature`. A triangular lattice fills the box, so the number of atoms has to be a number of columns times an even number of rows. Dividing the columns by the rows has to give something near `sqrt(3) / 2`, and `max_strain` is how far away it may sit, as a fraction of that ratio. At the default of 0.05 the counts up to 300 are 30, 56, 90, 120, 168, 224, 270 and 288, and any other count raises `ValueError` naming ones that fit. `timestep` is in seconds and applies to molecular dynamics only. `cut_off` is in Angstrom and defaults to 15 or half the box, whichever is smaller; it may not exceed half the box. `seed` seeds `simulation.rng`, which draws the initial velocities and the Monte Carlo moves.
 
 A configuration whose pair energy is not finite or exceeds ten $k_B T$ per atom is refused with `ValueError`, as is a potential whose energy at the cut-off is larger than $k_B T$.
